+++ app/scripts/controllers/permissions/index.js
@@ -1,12 +1,11 @@
-import JsonRpcEngine from 'json-rpc-engine'
-import asMiddleware from 'json-rpc-engine/src/asMiddleware'
-import ObservableStore from 'obs-store'
+import { JsonRpcEngine } from 'json-rpc-engine'
+import { ObservableStore } from '@metamask/obs-store'
 import log from 'loglevel'
 import { CapabilitiesController as RpcCap } from 'rpc-cap'
 import { ethErrors } from 'eth-json-rpc-errors'
 import { cloneDeep } from 'lodash'
 
-import createMethodMiddleware from './methodMiddleware'
+import createMethodMiddleware from './permissionsMethodMiddleware'
 import PermissionsLogController from './permissionsLog'
 
 // Methods that do not require any permissions to use:
@@ -90,7 +89,7 @@ export class PermissionsController {
       this.permissions, { origin },
     ))
 
-    return asMiddleware(engine)
+    return engine.asMiddleware(engine)
   }
 
   /**
@@ -102,11 +101,14 @@ export class PermissionsController {
    */
   getAccounts (origin) {
     // return new Promise((resolve, _) => {
-
     //   const req = { method: 'eth_accounts' }
     //   const res = {}
     //   this.permissions.providerMiddlewareFunction(
-    //     { origin }, req, res, () => {}, _end
+    //     { origin },
+    //     req,
+    //     res,
+    //     () => undefined,
+    //     _end,
     //   )
 
     //   function _end () {