+++ app/scripts/controllers/transactions/index.js
@@ -1,9 +1,9 @@
 import EventEmitter from 'safe-event-emitter'
-import ObservableStore from 'obs-store'
+import { ObservableStore } from '@metamask/obs-store'
 import ethUtil from 'ethereumjs-util'
 import Transaction from 'ethereumjs-tx'
 import EthQuery from 'ethjs-query'
-
+import { ethErrors } from 'eth-json-rpc-errors'
 import abi from 'human-standard-token-abi'
 import abiDecoder from 'abi-decoder'
 
@@ -14,14 +14,15 @@ import TxGasUtil from './tx-gas-utils'
 const PendingTransactionTracker = require('./pending-tx-tracker')
 import NonceTracker from 'nonce-tracker'
 import * as txUtils from './lib/util'
+import {
+  TRANSACTION_STATUSES,
+  TRANSACTION_TYPES,
+} from '../../../../shared/constants/transaction'
 import cleanErrorStack from '../../lib/cleanErrorStack'
 import log from 'loglevel'
 const recipientBlacklistChecker = require('./lib/recipient-blacklist-checker')
 const {
-  TRANSACTION_TYPE_CANCEL,
   TRANSACTION_TYPE_RETRY,
-  TRANSACTION_TYPE_STANDARD,
-  TRANSACTION_STATUS_APPROVED,
 } = require('./enums')
 
 const { hexToBn, bnToHex, BnMultiplyByFraction } = require('../../lib/util')
@@ -40,18 +41,17 @@ const { hexToBn, bnToHex, BnMultiplyByFraction } = require('../../lib/util')
     <br>- nonceTracker
       calculating nonces
 
-
   @class
-  @param {object} - opts
-  @param {object}  opts.initState - initial transaction list default is an empty array
-  @param {Object}  opts.networkStore - an observable store for network number
-  @param {Object}  opts.blockTracker - An instance of eth-blocktracker
-  @param {Object}  opts.provider - A network provider.
-  @param {Function}  opts.signTransaction - function the signs an ethereumjs-tx
-  @param {Function}  [opts.getGasPrice] - optional gas price calculator
-  @param {Function}  opts.signTransaction - ethTx signer that returns a rawTx
-  @param {Number}  [opts.txHistoryLimit] - number *optional* for limiting how many transactions are in state
-  @param {Object}  opts.preferencesStore
+  @param {Object} opts
+  @param {Object} opts.initState - initial transaction list default is an empty array
+  @param {Object} opts.networkStore - an observable store for network number
+  @param {Object} opts.blockTracker - An instance of eth-blocktracker
+  @param {Object} opts.provider - A network provider.
+  @param {Function} opts.signTransaction - function the signs an ethereumjs-tx
+  @param {Object} opts.getPermittedAccounts - get accounts that an origin has permissions for
+  @param {Function} opts.signTransaction - ethTx signer that returns a rawTx
+  @param {number} [opts.txHistoryLimit] - number *optional* for limiting how many transactions are in state
+  @param {Object} opts.preferencesStore
 */
 
 class TransactionController extends EventEmitter {
@@ -104,7 +104,12 @@ class TransactionController extends EventEmitter {
     this._updatePendingTxsAfterFirstBlock()
   }
 
-  /** @returns {number} the chainId*/
+  /**
+   * Gets the current chainId in the network store as a number, returning 0 if
+   * the chainId parses to NaN.
+   *
+   * @returns {number} The numerical chainId.
+   */
   getChainId () {
     const networkState = this.networkStore.getState()
     const getChainId = parseInt(networkState)
@@ -115,10 +120,10 @@ class TransactionController extends EventEmitter {
     }
   }
 
-/**
+  /**
   Adds a tx to the txlist
   @emits ${txMeta.id}:unapproved
-*/
+  */
   addTx (txMeta) {
     this.txStateManager.addTx(txMeta)
     this.emit(`${txMeta.id}:unapproved`, txMeta)
@@ -133,42 +138,66 @@ class TransactionController extends EventEmitter {
   }
 
   /**
-  add a new unapproved transaction to the pipeline
-
-  @returns {Promise<string>} the hash of the transaction after being submitted to the network
-  @param txParams {object} - txParams for the transaction
-  @param opts {object} - with the key origin to put the origin on the txMeta
-  */
-
+   * Add a new unapproved transaction to the pipeline
+   *
+   * @returns {Promise<string>} the hash of the transaction after being submitted to the network
+   * @param {Object} txParams - txParams for the transaction
+   * @param {Object} opts - with the key origin to put the origin on the txMeta
+   */
   async newUnapprovedTransaction (txParams, opts = {}) {
-    log.debug(`MetaMaskController newUnapprovedTransaction ${JSON.stringify(txParams)}`)
-    const initialTxMeta = await this.addUnapprovedTransaction(txParams)
-    initialTxMeta.origin = opts.origin
-    this.txStateManager.updateTx(initialTxMeta, '#newUnapprovedTransaction - adding the origin')
+    log.debug(
+      `MetaMaskController newUnapprovedTransaction ${JSON.stringify(txParams)}`,
+    )
+
+    const initialTxMeta = await this.addUnapprovedTransaction(
+      txParams,
+      opts.origin,
+    )
+
     // listen for tx completion (success, fail)
     return new Promise((resolve, reject) => {
-      this.txStateManager.once(`${initialTxMeta.id}:finished`, (finishedTxMeta) => {
-        switch (finishedTxMeta.status) {
-          case 'submitted':
-            return resolve(finishedTxMeta.hash)
-          case 'rejected':
-            return reject(cleanErrorStack(new Error('Nifty Wallet Tx Signature: User denied transaction signature.')))
-          case 'failed':
-            return reject(cleanErrorStack(new Error(finishedTxMeta.err.message)))
-          default:
-            return reject(cleanErrorStack(new Error(`Nifty Wallet Tx Signature: Unknown problem: ${JSON.stringify(finishedTxMeta.txParams)}`)))
-        }
-      })
+      this.txStateManager.once(
+        `${initialTxMeta.id}:finished`,
+        (finishedTxMeta) => {
+          switch (finishedTxMeta.status) {
+            case TRANSACTION_STATUSES.SUBMITTED:
+              return resolve(finishedTxMeta.hash)
+            case TRANSACTION_STATUSES.REJECTED:
+              return reject(
+                cleanErrorStack(
+                  ethErrors.provider.userRejectedRequest(
+                    'MetaMask Tx Signature: User denied transaction signature.',
+                  ),
+                ),
+              )
+            case TRANSACTION_STATUSES.FAILED:
+              return reject(
+                cleanErrorStack(
+                  ethErrors.rpc.internal(finishedTxMeta.err.message),
+                ),
+              )
+            default:
+              return reject(
+                cleanErrorStack(
+                  ethErrors.rpc.internal(
+                    `MetaMask Tx Signature: Unknown problem: ${JSON.stringify(
+                      finishedTxMeta.txParams,
+                    )}`,
+                  ),
+                ),
+              )
+          }
+        },
+      )
     })
   }
 
   /**
-  Validates and generates a txMeta with defaults and puts it in txStateManager
-  store
-
-  @returns {txMeta}
-  */
-
+   * Validates and generates a txMeta with defaults and puts it in txStateManager
+   * store.
+   *
+   * @returns {txMeta}
+   */
   async addUnapprovedTransaction (txParams) {
     // validate
     const normalizedTxParams = txUtils.normalizeTxParams(txParams)
@@ -180,7 +209,7 @@ class TransactionController extends EventEmitter {
     // construct txMeta
     let txMeta = this.txStateManager.generateTxMeta({
       txParams: normalizedTxParams,
-      type: TRANSACTION_TYPE_STANDARD,
+      type: TRANSACTION_TYPES.STANDARD,
     })
     this.addTx(txMeta)
     this.emit('newUnapprovedTx', txMeta)
@@ -201,11 +230,12 @@ class TransactionController extends EventEmitter {
 
     return txMeta
   }
-/**
-  adds the tx gas defaults: gas && gasPrice
-  @param txMeta {Object} - the txMeta object
-  @returns {Promise<object>} resolves with txMeta
-*/
+
+    /**
+   * Adds the tx gas defaults: gas && gasPrice
+   * @param {Object} txMeta - the txMeta object
+   * @returns {Promise<object>} resolves with txMeta
+   */
   async addTxGasDefaults (txMeta) {
     const txParams = txMeta.txParams
     // ensure value
@@ -247,7 +277,7 @@ class TransactionController extends EventEmitter {
    * new transaction contains the same nonce as the previous, is a basic ETH transfer of 0x value to
    * the sender's address, and has a higher gasPrice than that of the previous transaction.
    * @param {number} originalTxId - the id of the txMeta that you want to attempt to cancel
-   * @param {string=} customGasPrice - the hex value to use for the cancel transaction
+   * @param {string} [customGasPrice] - the hex value to use for the cancel transaction
    * @returns {txMeta}
    */
   async createCancelTransaction (originalTxId, customGasPrice) {
@@ -267,8 +297,8 @@ class TransactionController extends EventEmitter {
       },
       lastGasPrice,
       loadingDefaults: false,
-      status: TRANSACTION_STATUS_APPROVED,
-      type: TRANSACTION_TYPE_CANCEL,
+      status: TRANSACTION_STATUSES.APPROVED,
+      type: TRANSACTION_TYPES.CANCEL,
     })
 
     this.addTx(newTxMeta)
@@ -278,7 +308,7 @@ class TransactionController extends EventEmitter {
 
   /**
   updates the txMeta in the txStateManager
-  @param txMeta {Object} - the updated txMeta
+  @param {Object} txMeta - the updated txMeta
   */
   async updateTransaction (txMeta) {
     this.txStateManager.updateTx(txMeta, 'confTx: user updated transaction')
@@ -286,7 +316,7 @@ class TransactionController extends EventEmitter {
 
   /**
   updates and approves the transaction
-  @param txMeta {Object}
+  @param {Object} txMeta
   */
   async updateAndApproveTransaction (txMeta) {
     this.txStateManager.updateTx(txMeta, 'confTx: user approved transaction')
@@ -300,7 +330,7 @@ class TransactionController extends EventEmitter {
   signs the transaction
   publishes the transaction
   if any of these steps fails the tx status will be set to failed
-    @param txId {number} - the tx's Id
+    @param {number} txId - the tx's Id
   */
   async approveTransaction (txId, customNonce) {
     let nonceLock
@@ -338,10 +368,11 @@ class TransactionController extends EventEmitter {
       throw err
     }
   }
-  /**
+
+    /**
     adds the chain id and signs the transaction and set the status to signed
-    @param txId {number} - the tx's Id
-    @returns - rawTx {string}
+    @param {number} txId - the tx's Id
+    @returns {string} rawTx
   */
   async signTransaction (txId) {
     const txMeta = this.txStateManager.getTx(txId)
@@ -360,8 +391,8 @@ class TransactionController extends EventEmitter {
 
   /**
     publishes the raw tx and sets the txMeta to submitted
-    @param txId {number} - the tx's Id
-    @param rawTx {string} - the hex string of the serialized signed transaction
+    @param {number} txId - the tx's Id
+    @param {string} rawTx - the hex string of the serialized signed transaction
     @returns {Promise<void>}
   */
   async publishTransaction (txId, rawTx) {
@@ -413,7 +444,7 @@ class TransactionController extends EventEmitter {
 
   /**
     Convenience method for the ui thats sets the transaction to rejected
-    @param txId {number} - the tx's Id
+    @param {number} txId - the tx's Id
     @returns {Promise<void>}
   */
   async cancelTransaction (txId) {
@@ -422,8 +453,8 @@ class TransactionController extends EventEmitter {
 
   /**
     Sets the txHas on the txMeta
-    @param txId {number} - the tx's Id
-    @param txHash {string} - the hash for the txMeta
+    @param {number} txId - the tx's Id
+    @param {string} txHash - the hash for the txMeta
   */
   setTxHash (txId, txHash) {
     // Add the tx hash to the persisted meta-tx object
@@ -469,57 +500,96 @@ class TransactionController extends EventEmitter {
   */
 
   _onBootCleanUp () {
-    this.txStateManager.getFilteredTxList({
-      status: 'unapproved',
+  this.txStateManager
+    .getFilteredTxList({
+      status: TRANSACTION_STATUSES.UNAPPROVED,
       loadingDefaults: true,
-    }).forEach((tx) => {
+    })
+    .forEach((tx) => {
       this.addTxGasDefaults(tx)
-      .then((txMeta) => {
-        txMeta.loadingDefaults = false
-        this.txStateManager.updateTx(txMeta, 'transactions: gas estimation for tx on boot')
-      }).catch((error) => {
-        this.txStateManager.setTxStatusFailed(tx.id, error)
-      })
+        .then((txMeta) => {
+          txMeta.loadingDefaults = false
+          this.txStateManager.updateTx(
+            txMeta,
+            'transactions: gas estimation for tx on boot',
+          )
+        })
+        .catch((error) => {
+          const txMeta = this.txStateManager.getTx(tx.id)
+          txMeta.loadingDefaults = false
+          this.txStateManager.updateTx(
+            txMeta,
+            'failed to estimate gas during boot cleanup.',
+          )
+          this.txStateManager.setTxStatusFailed(txMeta.id, error)
+        })
     })
 
-    this.txStateManager.getFilteredTxList({
-      status: TRANSACTION_STATUS_APPROVED,
-    }).forEach((txMeta) => {
-      const txSignError = new Error('Transaction found as "approved" during boot - possibly stuck during signing')
+  this.txStateManager
+    .getFilteredTxList({
+      status: TRANSACTION_STATUSES.APPROVED,
+    })
+    .forEach((txMeta) => {
+      const txSignError = new Error(
+        'Transaction found as "approved" during boot - possibly stuck during signing',
+      )
       this.txStateManager.setTxStatusFailed(txMeta.id, txSignError)
     })
-  }
+}
 
   /**
     is called in constructor applies the listeners for pendingTxTracker txStateManager
     and blockTracker
   */
-  _setupListeners () {
-    this.txStateManager.on('tx:status-update', this.emit.bind(this, 'tx:status-update'))
-    this._setupBlockTrackerListener()
-    this.pendingTxTracker.on('tx:warning', (txMeta) => {
-      this.txStateManager.updateTx(txMeta, 'transactions/pending-tx-tracker#event: tx:warning')
-    })
-    this.pendingTxTracker.on('tx:failed', this.txStateManager.setTxStatusFailed.bind(this.txStateManager))
-    this.pendingTxTracker.on('tx:confirmed', (txId) => this.confirmTransaction(txId))
-    this.pendingTxTracker.on('tx:block-update', (txMeta, latestBlockNumber) => {
-      if (!txMeta.firstRetryBlockNumber) {
-        txMeta.firstRetryBlockNumber = latestBlockNumber
-        this.txStateManager.updateTx(txMeta, 'transactions/pending-tx-tracker#event: tx:block-update')
-      }
-    })
-    this.pendingTxTracker.on('tx:retry', (txMeta) => {
-      if (!('retryCount' in txMeta)) txMeta.retryCount = 0
-      txMeta.retryCount++
-      this.txStateManager.updateTx(txMeta, 'transactions/pending-tx-tracker#event: tx:retry')
-    })
-  }
+ _setupListeners () {
+  this.txStateManager.on(
+    'tx:status-update',
+    this.emit.bind(this, 'tx:status-update'),
+  )
+  this._setupBlockTrackerListener()
+  this.pendingTxTracker.on('tx:warning', (txMeta) => {
+    this.txStateManager.updateTx(
+      txMeta,
+      'transactions/pending-tx-tracker#event: tx:warning',
+    )
+  })
+  this.pendingTxTracker.on(
+    'tx:failed',
+    this.txStateManager.setTxStatusFailed.bind(this.txStateManager),
+  )
+  this.pendingTxTracker.on('tx:confirmed', (txId, transactionReceipt) =>
+    this.confirmTransaction(txId, transactionReceipt),
+  )
+  this.pendingTxTracker.on(
+    'tx:dropped',
+    this.txStateManager.setTxStatusDropped.bind(this.txStateManager),
+  )
+  this.pendingTxTracker.on('tx:block-update', (txMeta, latestBlockNumber) => {
+    if (!txMeta.firstRetryBlockNumber) {
+      txMeta.firstRetryBlockNumber = latestBlockNumber
+      this.txStateManager.updateTx(
+        txMeta,
+        'transactions/pending-tx-tracker#event: tx:block-update',
+      )
+    }
+  })
+  this.pendingTxTracker.on('tx:retry', (txMeta) => {
+    if (!('retryCount' in txMeta)) {
+      txMeta.retryCount = 0
+    }
+    txMeta.retryCount += 1
+    this.txStateManager.updateTx(
+      txMeta,
+      'transactions/pending-tx-tracker#event: tx:retry',
+    )
+  })
+}
 
   /**
     Sets other txMeta statuses to dropped if the txMeta that has been confirmed has other transactions
     in the list have the same nonce
 
-    @param txId {Number} - the txId of the transaction that has been confirmed in a block
+    @param {number} txId - the txId of the transaction that has been confirmed in a block
   */
   _markNonceDuplicatesDropped (txId) {
     // get the confirmed transactions nonce and from address
@@ -540,8 +610,7 @@ class TransactionController extends EventEmitter {
   _setupBlockTrackerListener () {
     let listenersAreActive = false
     const latestBlockHandler = this._onLatestBlock.bind(this)
-    const blockTracker = this.blockTracker
-    const txStateManager = this.txStateManager
+    const { blockTracker, txStateManager } = this
 
     txStateManager.on('tx:status-update', updateSubscription)
     updateSubscription()