+++ app/scripts/lib/personal-message-manager.js
@@ -1,11 +1,13 @@
 import EventEmitter from 'events'
-import ObservableStore from 'obs-store'
+import { ObservableStore } from '@metamask/obs-store'
 import ethUtil from 'ethereumjs-util'
 import { ethErrors } from 'eth-json-rpc-errors'
-import createId from './random-id'
-
-const hexRe = /^[0-9A-Fa-f]+$/g
 import log from 'loglevel'
+import { addHexPrefix } from './util'
+import createId from './random-id'
+import { MESSAGE_TYPE } from './enums'
+
+const hexRe = /^[0-9A-Fa-f]+$/gu
 
 /**
  * Represents, and contains data about, an 'personal_sign' type signature request. These are created when a
@@ -31,11 +33,10 @@ export default class PersonalMessageManager extends EventEmitter {
    * Controller in charge of managing - storing, adding, removing, updating - PersonalMessage.
    *
    * @typedef {Object} PersonalMessageManager
-   * @param {Object} opts @deprecated
-   * @property {Object} memStore The observable store where PersonalMessage are saved with persistance.
+   * @property {Object} memStore The observable store where PersonalMessage are saved.
    * @property {Object} memStore.unapprovedPersonalMsgs A collection of all PersonalMessages in the 'unapproved' state
    * @property {number} memStore.unapprovedPersonalMsgCount The count of all PersonalMessages in this.memStore.unapprobedMsgs
-   * @property {array} messages Holds all messages that have been created by this PersonalMessageManager
+   * @property {Array} messages Holds all messages that have been created by this PersonalMessageManager
    *
    */
   constructor () {
@@ -50,7 +51,7 @@ export default class PersonalMessageManager extends EventEmitter {
   /**
    * A getter for the number of 'unapproved' PersonalMessages in this.messages
    *
-   * @returns {number} - The number of 'unapproved' PersonalMessages in this.messages
+   * @returns {number} The number of 'unapproved' PersonalMessages in this.messages
    *
    */
   get unapprovedPersonalMsgCount () {
@@ -60,14 +61,16 @@ export default class PersonalMessageManager extends EventEmitter {
   /**
    * A getter for the 'unapproved' PersonalMessages in this.messages
    *
-   * @returns {Object} - An index of PersonalMessage ids to PersonalMessages, for all 'unapproved' PersonalMessages in
+   * @returns {Object} An index of PersonalMessage ids to PersonalMessages, for all 'unapproved' PersonalMessages in
    * this.messages
    *
    */
   getUnapprovedMsgs () {
-    return this.messages.filter((msg) => msg.status === 'unapproved')
+    return this.messages
+      .filter((msg) => msg.status === 'unapproved')
       .reduce((result, msg) => {
-        result[msg.id] = msg; return result
+        result[msg.id] = msg
+        return result
       }, {})
   }
 
@@ -77,24 +80,37 @@ export default class PersonalMessageManager extends EventEmitter {
    * this.memStore.
    *
    * @param {Object} msgParams - The params for the eth_sign call to be made after the message is approved.
-   * @param {Object} req (optional) The original request object possibly containing the origin
-   * @returns {promise} - When the message has been signed or rejected
+   * @param {Object} [req] - The original request object possibly containing the origin
+   * @returns {promise} When the message has been signed or rejected
    *
    */
   addUnapprovedMessageAsync (msgParams, req) {
     return new Promise((resolve, reject) => {
       if (!msgParams.from) {
         reject(new Error('Nifty Wallet Message Signature: from field is required.'))
+        return
       }
       const msgId = this.addUnapprovedMessage(msgParams, req)
       this.once(`${msgId}:finished`, (data) => {
         switch (data.status) {
           case 'signed':
-            return resolve(data.rawSig)
+            resolve(data.rawSig)
+            return
           case 'rejected':
-            return reject(ethErrors.provider.userRejectedRequest('Nifty Wallet Message Signature: User denied message signature.'))
+            reject(
+              ethErrors.provider.userRejectedRequest(
+                'Nifty Wallet Message Signature: User denied message signature.',
+              ),
+            )
+            return
           default:
-            return reject(new Error(`Nifty Wallet Message Signature: Unknown problem: ${JSON.stringify(msgParams)}`))
+            reject(
+              new Error(
+                `Nifty Wallet Message Signature: Unknown problem: ${JSON.stringify(
+                  msgParams,
+                )}`,
+              ),
+            )
         }
       })
     })
@@ -106,26 +122,30 @@ export default class PersonalMessageManager extends EventEmitter {
    * this.memStore.
    *
    * @param {Object} msgParams - The params for the eth_sign call to be made after the message is approved.
-   * @param {Object} req (optional) The original request object possibly containing the origin
-   * @returns {number} - The id of the newly created PersonalMessage.
+   * @param {Object} [req] - The original request object possibly containing the origin
+   * @returns {number} The id of the newly created PersonalMessage.
    *
    */
   addUnapprovedMessage (msgParams, req) {
-    log.debug(`PersonalMessageManager addUnapprovedMessage: ${JSON.stringify(msgParams)}`)
+    log.debug(
+      `PersonalMessageManager addUnapprovedMessage: ${JSON.stringify(
+        msgParams,
+      )}`,
+    )
     // add origin from request
     if (req) {
       msgParams.origin = req.origin
     }
     msgParams.data = this.normalizeMsgData(msgParams.data)
     // create txData obj with parameters and meta data
-    const time = (new Date()).getTime()
+    const time = new Date().getTime()
     const msgId = createId()
     const msgData = {
       id: msgId,
-      msgParams: msgParams,
-      time: time,
+      msgParams,
+      time,
       status: 'unapproved',
-      type: 'personal_sign',
+      type: MESSAGE_TYPE.PERSONAL_SIGN,
     }
     this.addMsg(msgData)
 
@@ -150,7 +170,7 @@ export default class PersonalMessageManager extends EventEmitter {
    * Returns a specified PersonalMessage.
    *
    * @param {number} msgId - The id of the PersonalMessage to get
-   * @returns {PersonalMessage|undefined} - The PersonalMessage with the id that matches the passed msgId, or undefined
+   * @returns {PersonalMessage|undefined} The PersonalMessage with the id that matches the passed msgId, or undefined
    * if no PersonalMessage has that id.
    *
    */
@@ -164,7 +184,7 @@ export default class PersonalMessageManager extends EventEmitter {
    *
    * @param {Object} msgParams - The msgParams to be used when eth_sign is called, plus data added by MetaMask.
    * @param {Object} msgParams.metamaskId Added to msgParams for tracking and identification within MetaMask.
-   * @returns {Promise<object>} - Promises the msgParams object with metamaskId removed.
+   * @returns {Promise<object>} Promises the msgParams object with metamaskId removed.
    *
    */
   approveMessage (msgParams) {
@@ -201,7 +221,7 @@ export default class PersonalMessageManager extends EventEmitter {
    * Removes the metamaskId property from passed msgParams and returns a promise which resolves the updated msgParams
    *
    * @param {Object} msgParams - The msgParams to modify
-   * @returns {Promise<object>} - Promises the msgParams with the metamaskId property removed
+   * @returns {Promise<object>} Promises the msgParams with the metamaskId property removed
    *
    */
   prepMsgForSigning (msgParams) {
@@ -235,7 +255,9 @@ export default class PersonalMessageManager extends EventEmitter {
   _setMsgStatus (msgId, status) {
     const msg = this.getMsg(msgId)
     if (!msg) {
-      throw new Error(`PersonalMessageManager - Message not found for id: "${msgId}".`)
+      throw new Error(
+        `PersonalMessageManager - Message not found for id: "${msgId}".`,
+      )
     }
     msg.status = status
     this._updateMsg(msg)
@@ -271,8 +293,12 @@ export default class PersonalMessageManager extends EventEmitter {
    */
   _saveMsgList () {
     const unapprovedPersonalMsgs = this.getUnapprovedMsgs()
-    const unapprovedPersonalMsgCount = Object.keys(unapprovedPersonalMsgs).length
-    this.memStore.updateState({ unapprovedPersonalMsgs, unapprovedPersonalMsgCount })
+    const unapprovedPersonalMsgCount = Object.keys(unapprovedPersonalMsgs)
+      .length
+    this.memStore.updateState({
+      unapprovedPersonalMsgs,
+      unapprovedPersonalMsgCount,
+    })
     this.emit('updateBadge')
   }
 
@@ -280,14 +306,14 @@ export default class PersonalMessageManager extends EventEmitter {
    * A helper function that converts raw buffer data to a hex, or just returns the data if it is already formatted as a hex.
    *
    * @param {any} data - The buffer data to convert to a hex
-   * @returns {string} - A hex string conversion of the buffer data
+   * @returns {string} A hex string conversion of the buffer data
    *
    */
   normalizeMsgData (data) {
     try {
       const stripped = ethUtil.stripHexPrefix(data)
       if (stripped.match(hexRe)) {
-        return ethUtil.addHexPrefix(stripped)
+        return addHexPrefix(stripped)
       }
     } catch (e) {
       log.debug(`Message was not hex encoded, interpreting as utf8.`)
@@ -295,5 +321,4 @@ export default class PersonalMessageManager extends EventEmitter {
 
     return ethUtil.bufferToHex(Buffer.from(data, 'utf8'))
   }
-
 }