+++ test/unit/app/controllers/detect-tokens-test.js
@@ -25,6 +25,8 @@ describe('DetectTokensController', () => {
 
     keyringMemStore = new ObservableStore({ isUnlocked: false})
     network = new NetworkController()
+    network.setEthMainnetRPCEndpoint('foo')
+    network.setInfuraProjectId('foo')
     preferences = new PreferencesController({ network })
     controller = new DetectTokensController({ preferences: preferences, network: network, keyringMemStore: keyringMemStore })