+++ ui/app/token-util.js
@@ -1,6 +1,6 @@
 import log from 'loglevel'
 import BigNumber from 'bignumber.js'
-import contractMapETH from 'eth-contract-metadata'
+import contractMapETH from '@metamask/contract-metadata'
 import contractMapPOA from 'poa-contract-metadata'
 import contractMapRSK from '@rsksmart/rsk-contract-metadata'
 import contractMapRSKTest from '@rsksmart/rsk-testnet-contract-metadata'